+++ packages/plugin-provider/src/omuchatprovider/errors.py
@@ -2,9 +2,3 @@ class ProviderError(Exception):
     """Base class for provider errors."""
 
     pass
-
-
-class ProviderFailed(ProviderError):
-    """Raised when a provider fails to update a channel."""
-
-    pass
