+++ packages/plugin-provider/src/omuchatprovider/services/youtube/__init__.py
@@ -1,3 +1,5 @@
 from .youtube import YoutubeService
 
-__all__ = ["YoutubeService"]
+__all__ = [
+    "YoutubeService",
+]
