+++ packages/plugin-provider/src/omuchatprovider/services/youtube/types/__init__.py
@@ -1,9 +1,11 @@
 from .live_chat import live_chat
+from .updated_metadata import updated_metadata
 from .ytcfg import ytcfg
 from .ytinitialdata import ytinitialdata
 
 __all__ = [
     "live_chat",
+    "updated_metadata",
     "ytcfg",
     "ytinitialdata",
 ]
