+++ packages/plugin-provider/src/omuchatprovider/services/youtube/types/live_chat.py
@@ -16,4 +16,4 @@ class ContinuationContents(TypedDict):
 
 
 class live_chat(YoutubeResponse):
-    trackingParams: str
+    continuationContents: ContinuationContents
